Fix integer overflow in _TIFFPartialReadStripArray on 32-bit
MR !851 fixed an identical overflow at line 8248 by adding a
(uint64_t) cast to prevent 32-bit wraparound in the multiplication
of sizeofval * strile. The same function has the same unchecked
pattern at line 8256 where arraySize * sizeofval can wrap on
32-bit platforms where size_t is 32-bit.

Apply the same (uint64_t) cast to ensure the multiplication is
performed in 64-bit arithmetic before addition with nBaseOffset.

+++ libtiff/tif_dirread.c
@@ -8253,7 +8253,7 @@ static int _TIFFPartialReadStripArray(TIFF *tif, TIFFDirEntry *dirent,
         nOffsetEndPage += IO_CACHE_PAGE_SIZE;
 #undef IO_CACHE_PAGE_SIZE
 
-    nLastStripOffset = nBaseOffset + arraySize * sizeofval;
+    nLastStripOffset = nBaseOffset + (uint64_t)arraySize * sizeofval;
     if (nLastStripOffset < nOffsetEndPage)
         nOffsetEndPage = nLastStripOffset;
     if (nOffsetStartPage >= nOffsetEndPage)
